+++ features/Context/Page/Batch/Operation.php
@@ -19,6 +19,7 @@ class Operation extends Wizard
         'Change status (enable / disable)' => 'Batch ChangeStatus',
         'Edit attributes'                  => 'Batch EditCommonAttributes',
         'Change the family of products'    => 'Batch ChangeFamily',
+        'Add to groups'                    => 'Batch AddToGroups',
     );
 
     /**